@@ -29,31 +29,38 @@
         "postversion": "git push && git push --tags"
     },
     "dependencies": {
-        "@fortawesome/fontawesome-free": "^5.11.2",
-        "vue": "^2.6.10",
-        "vue-ads-form-builder": "^0.1.15",
-        "vue-ads-pagination": "^2.1.7",
-        "vue-json-excel": "^0.2.98"
+        "@fortawesome/fontawesome-svg-core": "^6.1.1",
+        "@fortawesome/free-solid-svg-icons": "^6.1.1",
+        "@fortawesome/vue-fontawesome": "^3.0.0-5",
+        "core-js": "^3.21.1",
+        "uuid": "^8.3.2",
+        "vue": "^3.2.32"
     },
     "devDependencies": {
-        "@fullhuman/postcss-purgecss": "^1.3.0",
-        "@vue/cli-plugin-babel": "^4.0.5",
-        "@vue/cli-plugin-eslint": "^4.0.5",
-        "@vue/cli-plugin-unit-jest": "^4.0.5",
-        "@vue/cli-service": "^4.5.12",
-        "@vue/eslint-config-standard": "^4.0.0",
-        "@vue/test-utils": "^1.0.0-beta.29",
-        "babel-core": "7.0.0-bridge.0",
-        "babel-eslint": "^10.0.3",
-        "babel-jest": "^24.9.0",
-        "eslint": "^4.19.1",
-        "eslint-config-ads": "^1.0.7",
-        "eslint-plugin-vue": "^6.0.1",
+        "@babel/eslint-parser": "^7.17.0",
+        "@fullhuman/postcss-purgecss": "^4.1.3",
+        "@vue/cli-plugin-babel": "^5.0.4",
+        "@vue/cli-plugin-eslint": "^5.0.4",
+        "@vue/cli-plugin-unit-jest": "^5.0.4",
+        "@vue/cli-service": "^5.0.4",
+        "@vue/compiler-sfc": "^3.2.33",
+        "@vue/test-utils": "^2.0.0-0",
+        "@vue/vue3-jest": "^27.0.0-alpha.1",
+        "babel-jest": "^27.5.1",
+        "eslint": "^8.13.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-jest": "^26.1.4",
+        "eslint-plugin-prettier": "^4.0.0",
+        "eslint-plugin-vue": "^8.6.0",
+        "jest": "^27.5.1",
+        "jest-serializer-vue": "^2.0.2",
+        "jest-transform-stub": "^2.0.0",
         "postcss-import": "^12.0.1",
         "postcss-url": "^8.0.0",
         "prettier-package-json": "^2.1.0",
-        "tailwindcss": "^1.1.3",
-        "vue-template-compiler": "^2.6.10"
+        "purgecss": "^4.1.3",
+        "tailwindcss": "^3.0.24",
+        "url-loader": "^4.1.1"
     },
     "keywords": [
         "component",
