@@ -1,8 +1,4 @@
-import Vue from 'vue';
-import App from './TableContainerApp';
+import { createApp } from "vue";
+import App from "./TableContainerApp";
 
-Vue.config.productionTip = false;
-
-new Vue({
-    render: h => h(App),
-}).$mount('#app');
+createApp(App).mount("#app");
