@@ -1,18 +1,24 @@
 module.exports = {
-    root: true,
+  root: true,
 
-    env: {
-        node: true,
-    },
+  env: {
+    browser: true,
+    es6: true,
+    node: true,
+  },
 
-    extends: [
-        'plugin:vue/essential',
-        '@vue/standard',
-        'eslint-config-ads',
-        'eslint-config-ads/vue',
-    ],
+  extends: [
+    "plugin:vue/vue3-essential",
+    "plugin:prettier/recommended",
+    "eslint:recommended",
+  ],
 
-    parserOptions: {
-        parser: 'babel-eslint',
-    },
+  parserOptions: {
+    parser: "@babel/eslint-parser",
+  },
+
+  rules: {
+    "no-console": process.env.NODE_ENV === "production" ? "error" : "off",
+    "no-debugger": process.env.NODE_ENV === "production" ? "error" : "off",
+  },
 };
