@@ -1,5 +1,3 @@
 module.exports = {
-  presets: [
-    '@vue/app'
-  ]
+  presets: ["@vue/cli-plugin-babel/preset"],
 }
