@@ -1,7 +1,7 @@
-import './assets/css/tailwind.css';
+import "./assets/css/tailwind.css";
 
-import VueAdsTableContainer from './components/TableContainer';
-import VueAdsTable from './components/Table';
+import VueAdsTableContainer from "./components/TableContainer";
+import VueAdsTable from "./components/Table";
 
 export default VueAdsTableContainer;
 export { VueAdsTable };
