@@ -1,56 +1,57 @@
 export default {
-    table: {
-        'vue-ads-border': true,
-        'vue-ads-w-full': true,
-    },
-    info: {
-        'vue-ads-text-center': true,
-        'vue-ads-py-6': true,
-        'vue-ads-text-sm': true,
-        'vue-ads-border-t': true,
-    },
-    group: {
-        'vue-ads-font-bold': true,
-        'vue-ads-border-b': true,
-        'vue-ads-italic': true,
-    },
-    selected: {
-        'vue-ads-bg-teal-100': true,
-        'vue-ads-bg-gray-100': false,
-    },
-    exactMatch: {
-        'vue-ads-bg-yellow-300': true,
-    },
-    selectedAndExactMatch: {
-        selectedAndExactMatch: true,
-    },
-    'all/': {
-        'hover:vue-ads-bg-gray-200': true,
-    },
-    'all/all': {
-        'vue-ads-px-4': true,
-        'vue-ads-py-2': true,
-        'vue-ads-text-sm': true,
-    },
-    'even/': {
-        'vue-ads-bg-gray-100': true,
-    },
-    'odd/': {
-        'vue-ads-bg-white': true,
-    },
-    '0/': {
-        'vue-ads-bg-gray-100': false,
-        'hover:vue-ads-bg-gray-200': false,
-    },
-    '0/all': {
-        'vue-ads-px-4': true,
-        'vue-ads-py-2': true,
-        'vue-ads-text-left': true,
-    },
-    '0_-1/': {
-        'vue-ads-border-b': true,
-    },
-    '/0_-1': {
-        'vue-ads-border-r': true,
-    },
+  table: {
+    "vue-ads-border": true,
+    "vue-ads-w-full": true,
+  },
+  info: {
+    "vue-ads-text-center": true,
+    "vue-ads-py-6": true,
+    "vue-ads-text-sm": true,
+    "vue-ads-border-t": true,
+  },
+  group: {
+    "vue-ads-font-bold": true,
+    "vue-ads-border-b": true,
+    "vue-ads-italic": true,
+  },
+  selected: {
+    "vue-ads-bg-teal-100": true,
+    "vue-ads-bg-gray-100": false,
+    "vue-ads-bg-white": false,
+  },
+  exactMatch: {
+    "vue-ads-bg-yellow-300": true,
+  },
+  selectedAndExactMatch: {
+    selectedAndExactMatch: true,
+  },
+  "all/": {
+    "hover:vue-ads-bg-gray-200": true,
+  },
+  "all/all": {
+    "vue-ads-px-4": true,
+    "vue-ads-py-2": true,
+    "vue-ads-text-sm": true,
+  },
+  "even/": {
+    "vue-ads-bg-gray-100": true,
+  },
+  "odd/": {
+    "vue-ads-bg-white": true,
+  },
+  "0/": {
+    "vue-ads-bg-gray-100": false,
+    "hover:vue-ads-bg-gray-200": false,
+  },
+  "0/all": {
+    "vue-ads-px-4": true,
+    "vue-ads-py-2": true,
+    "vue-ads-text-left": true,
+  },
+  "0_-1/": {
+    "vue-ads-border-b": true,
+  },
+  "/0_-1": {
+    "vue-ads-border-r": true,
+  },
 };
