@@ -1,11 +1,12 @@
 module.exports = {
-    env: {
-        jest: true,
-    },
-    rules: {
-        'no-new': 'off',
-        'no-unused-expressions': 'off',
-        'no-template-curly-in-string': 'off',
-        'import/no-extraneous-dependencies': 'off',
-    },
+  env: {
+    jest: true,
+  },
+  plugins: ["jest"],
+  rules: {
+    "no-new": "off",
+    "no-unused-expressions": "off",
+    "no-template-curly-in-string": "off",
+    "import/no-extraneous-dependencies": "off",
+  },
 };
